[Design] : 작성하기 버튼 위치 수정
