+++ client/src/pages/Community/Main.tsx
@@ -6,6 +6,8 @@ import { AiFillHeart } from 'react-icons/ai';
 import { Link } from 'react-router-dom';
 import styled from 'styled-components';
 
+import Swal from 'sweetalert2';
+import { FiChevronRight } from 'react-icons/fi';
 import Pagination from '../../Components/Community/Pagination';
 import SideBar from '../../Components/Community/SideBar';
 import Tags from '../../Components/Community/Tags';
@@ -22,20 +24,50 @@ const Explain = styled.div`
 	padding: 30px;
 	line-height: 1.5rem;
 
-	> h1 {
-		margin-top: 20px;
-		margin-bottom: 10px;
-	}
-
 	> div {
-		color: #595959;
-		font-size: 14px;
-		padding-bottom: 10px;
-		border-bottom: 1px solid rgb(214, 217, 219);
+		> h1 {
+			margin-top: 20px;
+			margin-bottom: 10px;
+		}
+
+		> div {
+			color: #595959;
+			font-size: 14px;
+			padding-bottom: 10px;
+			border-bottom: 1px solid rgb(214, 217, 219);
+			display: flex;
+			justify-content: space-between;
+			align-items: end;
+
+			> button {
+				padding-bottom: 10px;
+				margin-right: 20px;
+
+				> span {
+					margin-right: 5px;
+					display: flex;
+					align-items: center;
+
+					> p {
+						display: flex;
+						align-items: center;
+					}
+				}
+
+				&:hover {
+					color: #0db4f3;
+
+					.arrow {
+						transform: translateX(4px);
+						transition: transform 0.3s ease-in-out;
+					}
+				}
+			}
+		}
 	}
 `;
 
-const Body = styled.div`
+const Container = styled.div`
 	height: 1000px;
 	display: flex;
 
@@ -50,7 +82,8 @@ const Body = styled.div`
 	}
 `;
 
-const ContentContainer = styled.div`
+const Body = styled.div`
+	height: calc(100vh - 260px);
 	width: calc(100vw - 400px);
 	margin-right: 30px;
 	height: fit-content;
@@ -69,6 +102,7 @@ const Contentbody = styled.div`
 	&:hover {
 		color: #0db4f3;
 	}
+
 	> div:nth-child(1) {
 		display: flex;
 		flex-direction: column;
@@ -90,6 +124,7 @@ const Contentbody = styled.div`
 
 const Header = styled.div`
 	padding: 5px;
+
 	> div {
 		display: flex;
 		-webkit-text-stroke: 0.4px black;
@@ -126,17 +161,9 @@ const Info = styled.div`
 	}
 `;
 
-const PaginationContainer = styled.div`
-	margin-top: 10px;
-	height: 32px;
-	display: flex;
-	justify-content: center;
-	align-items: center;
-	width: 100%;
-`;
-
 const TagContainer = styled.div`
 	height: 100%;
+	margin-top: 55px;
 	width: 230px;
 	margin-right: 20px;
 	display: flex;
@@ -148,6 +175,15 @@ const TagContainer = styled.div`
 		justify-content: center;
 		align-items: center;
 	}
+`;
+
+const PaginationContainer = styled.div`
+	margin-top: 10px;
+	height: 32px;
+	display: flex;
+	justify-content: center;
+	align-items: center;
+	width: 100%;
 `;
 
 function Main() {
@@ -175,6 +211,27 @@ function Main() {
 		url: '/posts',
 	});
 
+	const handleBtn = () => {
+		const Toast = Swal.mixin({
+			toast: true,
+			position: 'top',
+			showConfirmButton: false,
+			timer: 3000,
+			timerProgressBar: true,
+			didOpen: (toast: {
+				addEventListener: (arg0: string, arg1: any) => void;
+			}) => {
+				toast.addEventListener('mouseenter', Swal.stopTimer);
+				toast.addEventListener('mouseleave', Swal.resumeTimer);
+			},
+		});
+
+		Toast.fire({
+			icon: 'warning',
+			title: '로그인 상태가 아닙니다',
+		});
+	};
+
 	useEffect(() => {
 		if (response) {
 			setPosts(response);
@@ -188,20 +245,30 @@ function Main() {
 	return (
 		<div className="main">
 			<Explain>
-				<h1>커뮤니티</h1>
 				<div>
-					나와 같은 MBTI를 가진 사람들은 어떤 여행을 갔는지 궁금한가요? <br />
-					여행메이트가 없으세요? MBTI과몰입러라구요? 여러 잡담을 나누고 싶나요?
-					<br />
-					커뮤니티 각 탭을 누르며 둘러보세요!
+					<h1>커뮤니티</h1>
+					<div>
+						나와 같은 MBTI를 가진 사람들은 어떤 여행을 갔는지 궁금한가요? <br />
+						여행메이트가 없으세요? MBTI과몰입러라구요? 여러 잡담을 나누고
+						싶나요?
+						<br />
+						커뮤니티 각 탭을 누르며 둘러보세요!
+						<button onClick={handleBtn}>
+							<span>
+								작성하러 가기{' '}
+								<p className="arrow">
+									<FiChevronRight />
+								</p>
+							</span>
+						</button>
+					</div>
 				</div>
 			</Explain>
-
-			<Body>
+			<Container>
 				<SideBar />
 
 				<div>
-					<ContentContainer>
+					<Body>
 						{posts &&
 							posts.slice(startIdx, endIdx).map((el) => (
 								<Link to={`/community/${el.id}`}>
@@ -224,7 +291,6 @@ function Main() {
 													<p>{el.content}</p>
 												)}
 											</Header>
-
 											<Info>
 												<div>{el.nickName}</div>
 												<div>16:15</div>
@@ -242,7 +308,7 @@ function Main() {
 									</Contentbody>
 								</Link>
 							))}
-					</ContentContainer>
+					</Body>
 
 					<PaginationContainer>
 						<Pagination
@@ -251,15 +317,14 @@ function Main() {
 							totalPage={Math.ceil(posts.length / 8)}
 							totalCount={posts.length}
 							size={8}
-							pageCount={5}
+							pageCount={8}
 						/>
 					</PaginationContainer>
 				</div>
-
 				<TagContainer>
 					<Tags />
 				</TagContainer>
-			</Body>
+			</Container>
 		</div>
 	);
 }
